requirements.txt
requests
feedparser
markdownify
discord-webhook
python-dotenv